--- src/components/storeRegister/MyContext.jsx
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const MyContext = createContext();
-
-export default MyContext;
