--- src/components/storeRegister/style/CheckInfo.style.jsx
@@ -1,22 +1,0 @@
-import styled from 'styled-components';
-
-export const CheckDiv = styled.div`
-    margin-bottom: 20px;
-    width: 100%;
-    height: 70px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 100%;
-`;
-export const Title = styled.h3`
-    font-size: 30px;
-    margin-bottom: 10px;
-    color: var(--color-accent);
-`;
-
-export const Description = styled.p`
-    font-size: 16px;
-    color: gray;
-    margin-bottom: 10px;
-`;
